third_party/blink/web_tests/fast/js/constructor-length-expected.txt

This tests the length property of constructors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ArrayBuffer.length is 1
PASS AudioContext.length is 0
PASS Blob.length is 0
PASS CloseEvent.length is 1
PASS CustomEvent.length is 1
FAIL DOMFormData.length should be 0. Threw exception ReferenceError: DOMFormData is not defined
PASS DOMParser.length is 0
PASS DataView.length is 1
PASS ErrorEvent.length is 1
PASS Event.length is 1
PASS EventSource.length is 1
PASS Float32Array.length is 3
PASS Float64Array.length is 3
PASS FileReader.length is 0
FAIL FileReaderSync.length should be 0. Threw exception ReferenceError: FileReaderSync is not defined
PASS HashChangeEvent.length is 1
PASS Int16Array.length is 3
PASS Int32Array.length is 3
PASS Int8Array.length is 3
FAIL Intent.length should be 3. Threw exception ReferenceError: Intent is not defined
PASS MediaStream.length is 0
PASS MessageChannel.length is 0
PASS MessageEvent.length is 1
PASS PageTransitionEvent.length is 1
FAIL PeerConnection.length should be 2. Threw exception ReferenceError: PeerConnection is not defined
PASS PopStateEvent.length is 1
PASS ProgressEvent.length is 1
PASS ShadowRoot.length is 0
PASS SharedWorker.length is 1
PASS StorageEvent.length is 1
PASS TrackEvent.length is 1
PASS Uint16Array.length is 3
PASS Uint32Array.length is 3
PASS Uint8Array.length is 3
PASS Uint8ClampedArray.length is 3
PASS VTTCue.length is 3
PASS WebGLContextEvent.length is 1
PASS WebKitCSSMatrix.length is 0
PASS WebKitMutationObserver.length is 1
PASS WebSocket.length is 1
PASS Worker.length is 1
PASS URL.length is 1
PASS XMLHttpRequest.length is 0
PASS XMLSerializer.length is 0
PASS XPathEvaluator.length is 0
PASS XSLTProcessor.length is 0
PASS successfullyParsed is true

TEST COMPLETE
